update cmd struct
