+++ clihandler/cmd/submit/results.go
@@ -1,4 +1,4 @@
-package cmd
+package submit
 
 import (
 	"encoding/json"
@@ -54,53 +54,52 @@ func (resultsObject *ResultsObject) ListAllResources() (map[string]workloadinter
 	return map[string]workloadinterface.IMetadata{}, nil
 }
 
-var resultsCmd = &cobra.Command{
-	Use:   "results <json file>\nExample:\n$ kubescape submit results path/to/results.json --format-version v2",
-	Short: "Submit a pre scanned results file. The file must be in json format",
-	Long:  ``,
-	RunE: func(cmd *cobra.Command, args []string) error {
-		if len(args) == 0 {
-			return fmt.Errorf("missing results file")
-		}
+func getResultsCmd() *cobra.Command {
+	var resultsCmd = &cobra.Command{
+		Use:   "results <json file>\nExample:\n$ kubescape submit results path/to/results.json --format-version v2",
+		Short: "Submit a pre scanned results file. The file must be in json format",
+		Long:  ``,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) == 0 {
+				return fmt.Errorf("missing results file")
+			}
 
-		k8s := getKubernetesApi()
+			k8s := getKubernetesApi()
 
-		// get config
-		clusterConfig := getTenantConfig(submitInfo.Account, "", k8s)
-		if err := clusterConfig.SetTenant(); err != nil {
-			logger.L().Error("failed setting account ID", helpers.Error(err))
-		}
+			// get config
+			clusterConfig := getTenantConfig(submitInfo.Account, "", k8s)
+			if err := clusterConfig.SetTenant(); err != nil {
+				logger.L().Error("failed setting account ID", helpers.Error(err))
+			}
 
-		resultsObjects := NewResultsObject(clusterConfig.GetAccountID(), clusterConfig.GetClusterName(), args[0])
+			resultsObjects := NewResultsObject(clusterConfig.GetAccountID(), clusterConfig.GetClusterName(), args[0])
 
-		// submit resources
-		var r reporter.IReport
-		switch formatVersion {
-		case "v2":
-			r = reporterv2.NewReportEventReceiver(clusterConfig.GetConfigObj(), "")
-		default:
-			logger.L().Warning("Deprecated results version. run with '--format-version' flag", helpers.String("your version", formatVersion), helpers.String("latest version", "v2"))
-			r = reporterv1.NewReportEventReceiver(clusterConfig.GetConfigObj())
-		}
+			// submit resources
+			var r reporter.IReport
+			switch formatVersion {
+			case "v2":
+				r = reporterv2.NewReportEventReceiver(clusterConfig.GetConfigObj(), "")
+			default:
+				logger.L().Warning("Deprecated results version. run with '--format-version' flag", helpers.String("your version", formatVersion), helpers.String("latest version", "v2"))
+				r = reporterv1.NewReportEventReceiver(clusterConfig.GetConfigObj())
+			}
 
-		submitInterfaces := cliinterfaces.SubmitInterfaces{
-			ClusterConfig: clusterConfig,
-			SubmitObjects: resultsObjects,
-			Reporter:      r,
-		}
+			submitInterfaces := cliinterfaces.SubmitInterfaces{
+				ClusterConfig: clusterConfig,
+				SubmitObjects: resultsObjects,
+				Reporter:      r,
+			}
 
-		if err := clihandler.Submit(submitInterfaces); err != nil {
-			logger.L().Fatal(err.Error())
-		}
-		return nil
-	},
-}
-
-func init() {
-	submitCmd.AddCommand(resultsCmd)
+			if err := clihandler.Submit(submitInterfaces); err != nil {
+				logger.L().Fatal(err.Error())
+			}
+			return nil
+		},
+	}
 	resultsCmd.PersistentFlags().StringVar(&formatVersion, "format-version", "v1", "Output object can be differnet between versions, this is for maintaining backward and forward compatibility. Supported:'v1'/'v2'")
-}
 
+	return resultsCmd
+}
 func loadResultsFromFile(filePath string) ([]reporthandling.FrameworkReport, error) {
 	frameworkReports := []reporthandling.FrameworkReport{}
 	f, err := os.ReadFile(filePath)
