+++ core/kscore.go
@@ -1,0 +1,7 @@
+package core
+
+// Scan
+// List
+// Download
+// Config
+// Submit
